requirements.txt
playwright
python-dotenv
groq
pydantic
loguru
beautifulsoup4
pytest
pytest-asyncio
